Rename test_matmul_unary/binary to matmul_unary/binary_pattern.
