+++ tests/ap/make_axpr.sh
@@ -1,11 +1,5 @@
 #!/bin/bash
 
-TEST_FILENAME=${1:-"test_trivial_reduce"}
-#TEST_FILENAME=${1:-"test_matmul_unary"}
-
-echo "-- Write 'import ${TEST_FILENAME}' to __main__.py"
-echo "import ${TEST_FILENAME}" > __main__.py
-
 FILENAMES_ARRAY=(
     "index_code_gen_value_util"
     "index_drr_pass_util"
@@ -24,8 +18,8 @@ FILENAMES_ARRAY=(
     "abstract_drr"
     "ap_tpl_codegen"
     "matmul_variadic_tpl"
-    "test_matmul_unary"
-    "test_matmul_binary"
+    "matmul_unary_pattern"
+    "matmul_binary_pattern"
 )
 for filename in "${FILENAMES_ARRAY[@]}"
 do
